Share bisect model helper utilities
